Update to Silex 2 + Twig 2
* Added test for TwigProvider
* Updated to PHPUnit 6
* Added PHPCS
* Moved all documentation to README
* Moved to PSR-4 autoloading
* PHP 7.0 minimum requirement in composer.json
* Added PHP 7.1 testing
