+++ .travis.yml
@@ -1,9 +1,8 @@
 language: php
 
 php:
-  - 5.6
   - 7.0
-  - hhvm
-  
+  - 7.1
+
 before_script:
   - "sh bin/travis-init.sh"
