+++ LICENSE.md
@@ -1,4 +1,5 @@
-Copyright (c) 2011 Sven Eisenschmidt <sven.eisenschmidt@gmail.com>
+Copyright (c) 2017 Michael Heap <m@michaelheap.com>
+                   Sven Eisenschmidt <sven.eisenschmidt@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
