+++ src/MarkdownTwigExtension.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace SilexMarkdown\MarkdownExtension;
+namespace SilexMarkdown;
 
 use Knp\Bundle\MarkdownBundle\Parser\MarkdownParser;
 
@@ -17,17 +17,12 @@ class MarkdownTwigExtension extends \Twig_Extension
     public function getFilters()
     {
         return array(
-            'markdown' => new \Twig_Filter_Method($this, 'markdown', array('is_safe' => array('html'))),
+            'markdown' => new \Twig_SimpleFilter('markdown', array($this, 'markdown'), array('is_safe' => array('html'))),
         );
     }
 
     public function markdown($txt)
     {
         return $this->helper->transform($txt);
-    }
-
-    public function getName()
-    {
-        return 'markdown';
     }
 }
