+++ tests/SilexMarkdown/views/markdown_test.twig
@@ -1,0 +1,1 @@
+{{ input | markdown }}
