--- vendors.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-git clone https://github.com/fabpot/Twig 	 	      vendor/twig
-git clone https://github.com/knplabs/KnpMarkdownBundle   vendor/knplabs-markdown/Knp/Bundle/MarkdownBundle
